+++ src/encoded/visualization.py
@@ -1,12 +1,19 @@
 # import uuid
 #
 # from botocore.exceptions import ClientError
-# from copy import copy, deepcopy
+from copy import copy, deepcopy
+from datetime import datetime
+from dateutil.relativedelta import relativedelta
 # from dcicutils.misc_utils import print_error_message
 # from pyramid.httpexceptions import HTTPBadRequest
-# from pyramid.view import view_config
+from pyramid.view import view_config
 # from snovault import CONNECTION
-# from snovault.util import debug_log
+from snovault.util import debug_log
+from snovault.search.search import (
+    search as perform_search_request, 
+    make_search_subreq
+)
+from urllib.parse import urlencode
 #
 # from .types.base import SMAHTItem
 # from snovault.types.base import get_item_or_none
@@ -19,15 +26,196 @@
 # from snovault.util import make_s3_client
 #
 #
-# def includeme(config):
-#     config.add_route(
-#         'trace_workflow_runs',
-#         '/trace_workflow_run_steps/{file_uuid}/',
-#         traverse='/{file_uuid}'
-#     )
-#     config.add_route('get_higlass_viewconf', '/get_higlass_viewconf/')
-#     config.add_route('get_higlass_cohort_viewconf', '/get_higlass_cohort_viewconf/')
-#     config.scan(__name__)
+TERM_NAME_FOR_NO_VALUE  = "No value"
+
+# Common definition for aggregating all files **counts**.
+# This works four our ElasticSearch mapping though has some non-ideal-ities.
+# For example, we use "cardinality" instead of "value_count" agg (which would (more correctly) count duplicate files, etc.)
+SUM_FILES_AGGREGATION_DEFINITION = {
+    "total_files" : {
+        "cardinality" : {
+            "field" : "embedded.accession.raw",
+            "precision_threshold" : 10000
+        }
+    }
+}
+
+DATE_RANGE_PRESETS = ['all', 'custom', 'thismonth', 'previousmonth', 'last3months', 'last6months', 'last12months', 'thisyear', 'previousyear']
+
+def includeme(config):
+    # config.add_route(
+    #     'trace_workflow_runs',
+    #     '/trace_workflow_run_steps/{file_uuid}/',
+    #     traverse='/{file_uuid}'
+    # )
+    # config.add_route('get_higlass_viewconf', '/get_higlass_viewconf/')
+    # config.add_route('get_higlass_cohort_viewconf', '/get_higlass_cohort_viewconf/')
+    config.add_route('date_histogram_aggregations', '/date_histogram_aggregations/')
+    config.scan(__name__)
+
+
+@view_config(route_name='date_histogram_aggregations', request_method=['GET', 'POST'])
+@debug_log
+def date_histogram_aggregations(context, request):
+    '''PREDEFINED aggregations which run against type=File'''
+
+    # Defaults - may be overriden in URI params
+    date_histogram_fields    = ['file_status_tracking.uploading', 'file_status_tracking.uploaded', 'file_status_tracking.released']
+    group_by_fields          = [
+        'data_generation_summary.submission_centers', 'data_generation_summary.sequencing_center', 
+        'data_generation_summary.data_type', 'data_generation_summary.data_category', 'file_format.display_title',
+        'data_generation_summary.assays', 
+        'data_generation_summary.sequencing_platforms', 'dataset', 'software.display_title'
+        ]
+    date_histogram_intervals = ['weekly']
+
+    # Mapping of 'date_histogram_interval' options we accept to ElasticSearch interval vocab term.
+    interval_to_es_interval = {
+        'hourly'    : 'hour',
+        'daily'     : 'day',
+        'weekly'    : 'week',
+        'monthly'   : 'month',
+        'yearly'    : 'year'
+    }
+
+    try:
+        json_body = request.json_body
+        search_param_lists = json_body.get('search_query_params', {})
+    except Exception:
+        search_param_lists = request.GET.dict_of_lists()
+        if 'group_by' in search_param_lists:
+            group_by_fields = search_param_lists['group_by']
+            del search_param_lists['group_by'] # We don't wanna use it as search filter.
+            if len(group_by_fields) == 1 and group_by_fields[0] in ['None', 'null']:
+                group_by_fields = None
+        if 'date_histogram' in search_param_lists:
+            date_histogram_fields = search_param_lists['date_histogram']
+            del search_param_lists['date_histogram'] # We don't wanna use it as search filter.
+        if 'date_histogram_interval' in search_param_lists:
+            date_histogram_intervals = search_param_lists['date_histogram_interval']
+            for interval in date_histogram_intervals:
+                if interval not in interval_to_es_interval.keys():
+                    raise IndexError('"{}" is not one of daily, weekly, monthly, or yearly.'.format(interval))
+            del search_param_lists['date_histogram_interval'] # We don't wanna use it as search filter.
+        if 'date_range' in search_param_lists and len(search_param_lists['date_range']) > 0:
+            date_range = search_param_lists['date_range'][0]
+            date_from, date_to = convert_date_range(date_range)
+            if date_from is not None:
+                search_param_lists['{}.from'.format(date_histogram_fields[0])] = date_from.strftime("%Y-%m-%d")
+            if date_to is not None:
+                search_param_lists['{}.to'.format(date_histogram_fields[0])] = date_to.strftime("%Y-%m-%d")
+            del search_param_lists['date_range']
+        if not search_param_lists:
+            search_param_lists = {}
+
+    if 'File' in search_param_lists['type']:
+        # Add predefined sub-aggs to collect Exp and File counts from ExpSet items, in addition to getting own doc_count.
+
+        common_sub_agg = deepcopy(SUM_FILES_AGGREGATION_DEFINITION)
+
+        # Add on file_size_volume
+        for key_name in ['total_files']:
+            common_sub_agg[key_name + "_volume"] = {
+                "sum" : {
+                    "field" : common_sub_agg[key_name]["cardinality"]["field"].replace('.accession.raw', '.file_size')
+                }
+            }
+
+        if group_by_fields is not None:
+            group_by_agg_dict = {
+                group_by_field : {
+                    "terms" : {
+                        "field"     : "embedded." + group_by_field + ".raw",
+                        "missing"   : TERM_NAME_FOR_NO_VALUE,
+                        "size"      : 30
+                    },
+                    "aggs" : common_sub_agg
+                }
+                for group_by_field in group_by_fields if group_by_field is not None
+            }
+            histogram_sub_aggs = dict(common_sub_agg, **group_by_agg_dict)
+        else:
+            histogram_sub_aggs = common_sub_agg
+
+    else:
+        if group_by_fields is not None:
+            # Do simple date_histogram group_by sub agg, unless is set to 'None'
+            histogram_sub_aggs = {
+                group_by_field : {
+                    "terms" : {
+                        "field"     : "embedded." + group_by_field + ".raw",
+                        "missing"   : TERM_NAME_FOR_NO_VALUE,
+                        "size"      : 30
+                    }
+                }
+                for group_by_field in group_by_fields if group_by_field is not None
+            }
+        else:
+            histogram_sub_aggs = None
+
+    # Create an agg item for each interval in `date_histogram_intervals` x each date field in `date_histogram_fields`
+    # TODO: Figure out if we want to align these up instead of do each combination.
+    outer_date_histogram_agg = {}
+    for interval in date_histogram_intervals:
+        for dh_field in date_histogram_fields:
+            outer_date_histogram_agg[interval + '_interval_' + dh_field] = {
+                "date_histogram" : {
+                    "field": "embedded." + dh_field,
+                    "interval": interval_to_es_interval[interval],
+                    "format": "yyyy-MM-dd"
+                }
+            }
+            if histogram_sub_aggs:
+                outer_date_histogram_agg[interval + '_interval_' + dh_field]['aggs'] = histogram_sub_aggs
+
+    search_param_lists['limit'] = search_param_lists['from'] = [0]
+    subreq          = make_search_subreq(request, '{}?{}'.format('/search/', urlencode(search_param_lists, True)) )
+    search_result   = perform_search_request(None, subreq, custom_aggregations=outer_date_histogram_agg)
+
+    # remove unnecessary fields from result
+    for field_to_delete in ['@context', '@id', '@type', '@graph', 'title', 'filters', 'facets', 'sort', 'clear_filters', 'actions', 'columns']:
+        if search_result.get(field_to_delete) is None:
+            continue
+        del search_result[field_to_delete]
+
+    return search_result
+
+
+def convert_date_range(date_range_str):
+    
+    data_range_split = date_range_str.split('|')
+    preset = data_range_split[0]
+    if preset not in DATE_RANGE_PRESETS:
+        raise IndexError('"{}" is not one of {}.'.format(preset, ', '.join(DATE_RANGE_PRESETS)))
+    
+    today = datetime.today()
+    first_day_this_month = today.replace(day=1)
+    date_from, date_to = None, None
+
+    if preset == 'thismonth':
+        date_from = first_day_this_month
+    elif preset == 'previousmonth':
+        date_from = first_day_this_month - relativedelta(months=1)
+        date_to = first_day_this_month - relativedelta(days=1)
+    elif preset == 'last3months':
+        date_from = first_day_this_month - relativedelta(months=2)
+    elif preset == 'last6months':
+        date_from = first_day_this_month - relativedelta(months=5)
+    elif preset == 'last12months':
+        date_from = first_day_this_month - relativedelta(months=11)
+    elif preset == 'thisyear':
+        date_from = datetime(today.year, 1, 1)
+    elif preset == 'previousyear':
+        date_from = datetime(today.year - 1, 1, 1)
+        date_to = datetime(today.year - 1, 12, 31)
+    elif preset == 'custom':
+        if len(data_range_split) > 1 and data_range_split[1] and len(data_range_split[1]) == 10:
+            date_from = datetime.strptime(data_range_split[1], '%Y-%m-%d')
+        if len(data_range_split) > 2 and data_range_split[2] and len(data_range_split[2]) == 10:
+            date_to = datetime.strptime(data_range_split[2], '%Y-%m-%d')
+    
+    return [date_from, date_to]
+
 #
 #
 # # TODO: figure out how to make one of those cool /file/ACCESSION/@@download/-like URLs for this.